Componentize and organize.
