+++ packages/webapp/src/delegates/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./arrow-container";
-export * from "./delegate-funds-available";
 export * from "./level-heading";
 export * from "./my-delegation";
 export * from "./statuses";
