+++ packages/webapp/src/members/components/index.ts
@@ -1,3 +1,4 @@
+export * from "./funds-available-cta";
 export * from "./member-card";
 export * from "./member-chip";
 export * from "./member-collections";
@@ -6,3 +7,4 @@ export * from "./member-holo-card";
 export * from "./member-social-links";
 export * from "./members-grid";
 export * from "./token-balance";
+export * from "./withdraw-funds";
