+++ packages/webapp/src/pages/members/[id].tsx
@@ -14,7 +14,7 @@ import {
     MemberHoloCard,
     useMemberByAccountName,
 } from "members";
-import { DelegateFundsAvailable } from "delegates/components";
+import { FundsAvailableCTA } from "members/components";
 
 export const MemberPage = () => {
     const router = useRouter();
@@ -43,7 +43,7 @@ export const MemberPage = () => {
 
     return (
         <MemberPageContainer pageTitle={`${member.name}'s Profile`}>
-            <DelegateFundsAvailable account={member.account} />
+            <FundsAvailableCTA account={member.account} />
             <Container className="flex justify-center">
                 <MemberHoloCard member={member} className="max-w-xl" />
             </Container>
